fix package dependencies after renaming
