+++ org.abapgit.adt.backend/META-INF/MANIFEST.MF
@@ -21,6 +21,5 @@ Automatic-Module-Name: org.abapgit.adt
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Eclipse-ExtensibleAPI: true
-Export-Package: org.abapgit.adt,
- src.org.abapgit.adt
+Export-Package: org.abapgit.adt.backend
 Import-Package: com.sap.adt.project.ui.util
