+++ org.abapgit.adt.backend/src/org/abapgit/adt/backend/IRepository.java
@@ -1,4 +1,4 @@
-package org.abapgit.adt;
+package org.abapgit.adt.backend;
 
 public interface IRepository {
