+++ org.abapgit.adt.ui/META-INF/MANIFEST.MF
@@ -21,8 +21,8 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="[3.12.0,4.0.0)",
  com.sap.adt.projectexplorer.ui;bundle-version="[2.89.0,3.0.0)",
  com.sap.adt.communication;bundle-version="[2.89.1,3.0.0)",
  com.sap.adt.compatibility;bundle-version="[2.89.1,3.0.0)",
- org.abapgit.adt.backend;bundle-version="1.0.0",
- com.sap.adt.project.ui
+ com.sap.adt.project.ui,
+ org.abapgit.adt.backend;bundle-version="1.0.0"
 Automatic-Module-Name: org.abapgit.adt.ui
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
